pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Hexlet Page_Analyzer project"
authors = ["Dmitry Zhiryakov <d.a.zhiryakov@yandex.ru>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.5"
validators = "^0.20.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
